Make plugin disabled by default, with an option to enable it
Any workarounds for pathfinding issues are just that, workarounds. They
obscure potentially real issues in a park, like disconnected paths. I
still find this plugin useful in specific cases where pathfinding is
working really poorly for a lot of guests, but in general I don't think
it makes sense for this plugin to be enabled by default for all parks.